+++ stable_baselines/common/vec_env/dummy_vec_env.py
@@ -69,37 +69,22 @@ class DummyVecEnv(VecEnv):
     def _obs_from_buf(self):
         return dict_to_obs(self.observation_space, copy_obs_dict(self.buf_obs))
 
-    def env_method(self, method_name, *method_args, **method_kwargs):
-        """
-        Provides an interface to call arbitrary class methods of vectorized environments
-
-        :param method_name: (str) The name of the env class method to invoke
-        :param method_args: (tuple) Any positional arguments to provide in the call
-        :param method_kwargs: (dict) Any keyword arguments to provide in the call
-        :return: (list) List of items retured by the environment's method call
-        """
-        return [getattr(env_i, method_name)(*method_args, **method_kwargs) for env_i in self.envs]
-
-    def get_attr(self, attr_name):
-        """
-        Provides a mechanism for getting class attribues from vectorized environments
-
-        :param attr_name: (str) The name of the attribute whose value to return
-        :return: (list) List of values of 'attr_name' in all environments
-        """
-        return [getattr(env_i, attr_name) for env_i in self.envs]
+    def get_attr(self, attr_name, indices=None):
+        """Return attribute from vectorized environment (see base class)."""
+        target_envs = self._get_target_envs(indices)
+        return [getattr(env_i, attr_name) for env_i in target_envs]
 
     def set_attr(self, attr_name, value, indices=None):
-        """
-        Provides a mechanism for setting arbitrary class attributes inside vectorized environments
+        """Set attribute inside vectorized environments (see base class)."""
+        target_envs = self._get_target_envs(indices)
+        for env_i in target_envs:
+            setattr(env_i, attr_name, value)
 
-        :param attr_name: (str) Name of attribute to assign new value
-        :param value: (obj) Value to assign to 'attr_name'
-        :param indices: (list,int) Indices of envs to assign value
-        :return: (list) in case env access methods might return something, they will be returned in a list
-        """
-        if indices is None:
-            indices = range(len(self.envs))
-        elif isinstance(indices, int):
-            indices = [indices]
-        return [setattr(env_i, attr_name, value) for env_i in [self.envs[i] for i in indices]]
+    def env_method(self, method_name, *method_args, indices=None, **method_kwargs):
+        """Call instance methods of vectorized environments."""
+        target_envs = self._get_target_envs(indices)
+        return [getattr(env_i, method_name)(*method_args, **method_kwargs) for env_i in target_envs]
+
+    def _get_target_envs(self, indices):
+        indices = self._get_indices(indices)
+        return [self.envs[i] for i in indices]